pyproject.toml
[project]
name = "mrt-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.5",
]
